+++ crengine/src/lvfont/lvfontcache.cpp
@@ -63,7 +63,7 @@ LVFontCacheItem *LVFontCache::findFallback(lString8 face, int size) {
             best_index = i;
         }
     }
-    if (best_index <= 0)
+    if (best_index < 0 || best_match <= 0)
         return NULL;
     if (best_instance_match >= best_match)
         return _instance_list[best_instance_index];